Add features
- Add help menu
- Add question boxes
- Delete unnecessary code
- Change info in About

+++ ePy_GUI.py
@@ -15,11 +15,13 @@ from multiprocessing.dummy import Pool
 from decimal import *
 import math
 import sys
+import AboutGUI as ab
 
-### ePy_MC Mod ###
 
-def one_div(fact):
-    return Decimal(1)/fact
+##---ePy_MC Mod --------------------------------------------------------------##
+
+def one_div(num):
+    return Decimal(1)/num
 
 def e_cal(l, cores):
     
@@ -53,14 +55,9 @@ def e_cal(l, cores):
         temp_fact = i
         i += cores
         c += 1
-##        sys.stdout.write("\r%i loops passed.\n" % (c) ) #Print Loop status
-##        sys.stdout.flush()
-        #print i, "loops passed."
         if e == temp:
             break
         temp = e
-##    sys.stdout.write("\r%i loops passed.\n" % (c) )
-##    sys.stdout.flush()
     progressbar.step(-prog_stat)
     progressbar.grid(row=4, columnspan=2)
     root.update()
@@ -99,52 +96,49 @@ def e_cal_sc(l):
         temp_fact = i
         i += 1
         c += 1
-##        sys.stdout.write("\r%i loops passed.\n" % (c) ) #Print Loop status
-##        sys.stdout.flush()
-        #print i, "loops passed."
         if e == temp:
             break
         temp = e
-##    sys.stdout.write("\r%i loops passed.\n" % (c) )
-##    sys.stdout.flush()
     progressbar.step(-prog_stat)
     progressbar.grid(row=4, columnspan=2)
     root.update()
     return e
 
-###  ###
+##----------------------------------------------------------------------------##
 
 def p():
-    global E, TIMER_NUM
+    global E, TIMER_NUM, PROGRESSED
     ran = False
     try:
         dig = int(digit_select_box.get())
         B.config(state=DISABLED)
         ran = True
         if dig not in [100,500,1000,2000,5000,10000,20000,50000]:
-            tkMessageBox.showinfo( "Error", "Invalid digit number.")
+            tkMessageBox.showerror( "Error", "Invalid digit number.")
             B.config(state=NORMAL)
             return None
         else:
-            tkMessageBox.showinfo( "Warning", "Are you ready to calculate "+str(dig)+" digits of e?")
-            stop_b.config(state=NORMAL)
-            t = time.time()
-            if int(cores.get()) == 0:
+            if tkMessageBox.askyesno( "Warning", "Are you ready to calculate "+str(dig)+" digits of e?"):
+                stop_b.config(state=NORMAL)
                 t = time.time()
-                E = e_cal_sc(dig)
+                if int(cores.get()) == 0:
+                    t = time.time()
+                    E = e_cal_sc(dig)
+                else:
+                    t = time.time()
+                    E = e_cal(dig, multiprocessing.cpu_count())
+                if active_stat.get():
+                    TIMER_NUM = str(time.time()-t)
+                    TIMER.set(TIMER_NUM+" sec.")
+                else:
+                    TIMER.set("- sec.")
+                active_stat.set(True)
+                B.config(state=NORMAL)
             else:
-                t = time.time()
-                E = e_cal(dig, multiprocessing.cpu_count())
-            if active_stat.get():
-                TIMER_NUM = str(time.time()-t)
-                TIMER.set(TIMER_NUM+" sec.")
-            else:
-                TIMER.set("- sec.")
-            active_stat.set(True)
-            B.config(state=NORMAL)
+                B.config(state=NORMAL)
     except:
         if ran:
-            tkMessageBox.showinfo( "Error", "You had force stop or somthing else error.")
+            tkMessageBox.showerror( "Error", "You had force stop or somthing else error.")
             progressbar.step(-PROGRESSED)
             progressbar.grid(row=4, columnspan=2)
             root.update()
@@ -152,30 +146,38 @@ def p():
             tkMessageBox.showinfo( "Error", "Please select value!")
     B.config(state=NORMAL)
     stop_b.config(state=DISABLED)
+
 def force_stop():
     active_stat.set(False)
 
 def export():
     try:
         global E, TIMER_NUM
-        f = open("ePy_output.txt", "w")
-        f.write("ePy had burned your computer for %s sec. \n" % TIMER_NUM)
-        f.write("Here is your computer's value of e ("+str(len(str(E))-2)+" digits):\n")
-        for i in xrange(len(str(E))):
-            if i % 80 == 0:
-                f.write("\n")
-            f.write("%s" % (str(E)[i]))
-        f.close()
-        tkMessageBox.showinfo( "Operation Complete", "Exported file (ePy_output.txt) is in same folder of this program.")
+        if tkMessageBox.askyesno( "Export?", "Do you really want to export result file?\n (Will replace old file!)"):
+            f = open("ePy_output.txt", "w")
+            f.write("ePy had burned your computer for %s sec. \n" % TIMER_NUM)
+            f.write("Here is your computer's value of e ("+str(len(str(E))-2)+" digits):\n")
+            for i in xrange(len(str(E))):
+                if i % 80 == 0:
+                    f.write("\n")
+                f.write("%s" % (str(E)[i]))
+            f.close()
+            tkMessageBox.showinfo( "Operation Complete", "Exported file (ePy_output.txt) is in same folder of this program.")
     except:
-        tkMessageBox.showinfo( "Operation fail", "You haven't started it yet.")
+        tkMessageBox.showerror( "Error", "You haven't started it yet.")
+
 if __name__ == "__main__":
     root = Tk()
+    root.title("ePy bench")
+
+    #Global_Var#
     PROGRESSED = 0
     TIMER = StringVar()
     TIMER.set("N/A")
     E = 0
-    root.title("ePy bench")
+    #/Global_Var#
+
+    #GUI#
     active_stat = BooleanVar(root)
     active_stat.set(True)
 
@@ -184,21 +186,10 @@ if __name__ == "__main__":
     c.grid(row=1, columnspan=2)
 
     cores_l = Label(root, text="Number of digits:")
-    cores_l.grid(row=0, column=0, sticky="W")
+    cores_l.grid(row=0, column=0)
 
     digit_select_box = Spinbox(root, values=(100,500,1000,2000,5000,10000,20000,50000), width=8)
-    digit_select_box.grid(row=0, column=1, sticky="E")
-
-    ##Lb1 = Listbox(root)
-    ##Lb1.insert(1, "100")
-    ##Lb1.insert(2, "500")
-    ##Lb1.insert(3, "1000")
-    ##Lb1.insert(4, "2000")
-    ##Lb1.insert(5, "5000")
-    ##Lb1.insert(6, "10000")
-    ##Lb1.insert(7, "20000")
-    ##Lb1.insert(8, "50000")
-    ##Lb1.grid(row=1, column=0)
+    digit_select_box.grid(row=0, column=1)
 
     B = Button(root, text ="      Start      ", command = p)
     B.grid(row=2, column=0)
@@ -207,21 +198,31 @@ if __name__ == "__main__":
     stop_b.config(state=DISABLED)
     stop_b.grid(row=2, column=1, sticky='W')
 
-    progressbar = ttk.Progressbar(orient=HORIZONTAL, length=200, maximum=1.0, mode="determinate", variable=PROGRESSED)
+    progressbar = ttk.Progressbar(orient=HORIZONTAL, length=200,
+                                  maximum=1.0, mode="determinate", variable=PROGRESSED)
     progressbar.grid(row=4, columnspan=2)
 
 
     stat = Label(root, textvariable=TIMER)
     stat.grid(row=3, columnspan=2)
-    #MENUBAR
+    
+    #MENUBAR#
     menubar = Menu(root)
-    # create a pulldown menu, and add it to the menu bar
+
     filemenu = Menu(menubar, tearoff=0)
-    filemenu.add_command(label="Export last result", command=export)
+    filemenu.add_command(label="Export last test result", command=export)
     filemenu.add_separator()
     filemenu.add_command(label="Exit", command=root.quit)
     menubar.add_cascade(label="File", menu=filemenu)
+
+    filemenu = Menu(menubar, tearoff=0)
+    filemenu.add_command(label="About", command=ab.about_pop)
+    filemenu.add_command(label="Help", command=ab.help_pop)
+    menubar.add_cascade(label="Help", menu=filemenu)
+    
     root.config(menu=menubar)
     #/MENUBAR
+    
     root.mainloop()
+    #/GUI#
     exit()
